assignment08
Got some functionality. Had some problems with the navigator.geolocation position property. I did create the WeatherForcast Class but didn't get it implemented but if you could look at it so see if it is set up right I would appreciate the feedback.I did get the clock drawn and functioning. And I got the locator to allow.